+++ app/javascript/vue/tasks/nomenclature/match/components/Table/TableUnmatched.vue
@@ -1,0 +1,27 @@
+<template>
+  <table class="table-striped">
+    <thead>
+      <tr>
+        <th>Unmatched</th>
+      </tr>
+    </thead>
+    <tbody>
+      <tr
+        v-for="unmatched in list"
+        :key="unmatched"
+      >
+        <td>{{ unmatched }}</td>
+      </tr>
+    </tbody>
+  </table>
+</template>
+
+<script setup>
+
+defineProps({
+  list: {
+    type: Array,
+    required: true
+  }
+})
+</script>